requirements.txt
python-dotenv==1.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pandas==2.1.4
pytz==2023.3.post1
alembic==1.13.1
fastapi==0.104.1        
uvicorn[standard]==0.24.0.post1 
celery==5.3.6          
redis==5.0.1           